Split up conky.config.
Config is now split into 3 parts and merged dynamically:
* core config - this should always be the same for polycore
* wm config - window manager specific issues
* script config - config specific to this instance / theme

The script config is able to overwrite any of the other options
if desired.

The reason for this is two fold. 1 i'm sick of dealing with some
of the differences here, 2 as themes move more toward being lua
only like some of the examples i'd like to reduce the amount of
unneeded code in the main configuration file.